pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pymh"
version = "0.1.2"
description = "Monadic Harness — structured entropy management for long-horizon agent tasks"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
Homepage = "https://github.com/lulullama42/monadic-harness"
Repository = "https://github.com/lulullama42/monadic-harness"

[project.scripts]
pymh = "pymh.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/pymh"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"

[dependency-groups]
dev = [
    "hypothesis>=6.100",
    "pytest>=8.4.2",
    "ruff>=0.8",
]
